First pass at factoring out AWS file upload to helper.
This fails when trying to unzip and re-upload file
but it uploads the zip and downloads it.

You have to add AWS_PLATFORM_FILES_BUCKET to your shell environment
for this to work. The other AWS keys should already be there and/or
have defaults.
