+++ .env.example
@@ -1,10 +1,9 @@
-# AWS access
-AWS_REGION=us-east-2
-AWS_ACCESS_KEY_ID=yourAWSaccesskey
-AWS_SECRET_ACCESS_KEY=yourAWSsecretkey
-# AWS S3 bucket for Rise360
-AWS_S3_BUCKET=platform-dev-file-uploads
 APPLICATION_HOST=platformweb
+# AWS S3 bucket for Rise360 file uploads (and others). Should be set in your shell
+AWS_REGION=us-east-2
+#AWS_PLATFORM_FILES_BUCKET=the-bucket-name
+#AWS_ACCESS_KEY_ID=your_access_key
+#AWS_SECRET_ACCESS_KEY=your_secret_access_key
 BOOSTER_PORTAL_DOMAIN=boosterweb
 # Toggle this on to switch to the old CAS authentication agains the Join server
 #BZ_AUTH_SERVER=joinweb
