+++ config/secrets.yml
@@ -1,5 +1,9 @@
 development:
   application_host: <%= ENV.fetch('APPLICATION_HOST') { 'localhost' } %>
+  aws_files_bucket: <%= ENV['AWS_PLATFORM_FILES_BUCKET'] %>
+  aws_access_key: <%= ENV['AWS_ACCESS_KEY_ID'] %>
+  aws_secret_access_key: <%= ENV['AWS_SECRET_ACCESS_KEY'] %>
+  aws_region: <%= ENV.fetch('AWS_REGION') { 'us-east' } %>
   database_host: <%= ENV['DATABASE_HOST'] %>
   database_username: <%= ENV['DATABASE_USERNAME'] %>
   database_password: <%= ENV['DATABASE_PASSWORD'] %>
@@ -28,6 +32,10 @@ development:
 
 test:
   application_host: <%= ENV.fetch('APPLICATION_HOST') { 'localhost' } %>
+  aws_files_bucket: <%= ENV.fetch('AWS_PLATFORM_FILES_BUCKET') { 'your-bucket-name' } %>
+  aws_access_key: <%= ENV.fetch('AWS_ACCESS_KEY_ID') { 'AKIA55555555555555FM' } %>
+  aws_secret_access_key: <%= ENV.fetch('AWS_SECRET_ACCESS_KEY') { 'sXY55555555555555555555555555555555S/ZAu' } %>
+  aws_region: <%= ENV.fetch('AWS_REGION') { 'us-east' } %>
   database_host: <%= ENV['DATABASE_HOST'] %>
   database_username: <%= ENV['DATABASE_USERNAME'] %>
   database_password: <%= ENV['DATABASE_PASSWORD'] %>
@@ -55,6 +63,10 @@ test:
 
 staging:
   application_host: <%= ENV.fetch('APPLICATION_HOST') { nil } %>
+  aws_files_bucket: <%= ENV['AWS_PLATFORM_FILES_BUCKET'] %>
+  aws_access_key: <%= ENV['AWS_ACCESS_KEY_ID'] %>
+  aws_secret_access_key: <%= ENV['AWS_SECRET_ACCESS_KEY'] %>
+  aws_region: <%= ENV.fetch('AWS_REGION') { 'us-east' } %>
   database_host: <%= ENV['DATABASE_HOST'] %>
   database_username: <%= ENV['DATABASE_USERNAME'] %>
   database_password: <%= ENV['DATABASE_PASSWORD'] %>
@@ -82,6 +94,10 @@ staging:
 
 production:
   application_host: <%= ENV.fetch('APPLICATION_HOST') { nil } %>
+  aws_files_bucket: <%= ENV['AWS_PLATFORM_FILES_BUCKET'] %>
+  aws_access_key: <%= ENV['AWS_ACCESS_KEY_ID'] %>
+  aws_secret_access_key: <%= ENV['AWS_SECRET_ACCESS_KEY'] %>
+  aws_region: <%= ENV.fetch('AWS_REGION') { 'us-east' } %>
   database_host: <%= ENV['DATABASE_HOST'] %>
   database_username: <%= ENV['DATABASE_USERNAME'] %>
   database_password: <%= ENV['DATABASE_PASSWORD'] %>
