+++ docker-compose.yml
@@ -38,6 +38,11 @@ services:
       SALESFORCE_PLATFORM_USERNAME: ${SALESFORCE_PLATFORM_USERNAME}
       SALESFORCE_PLATFORM_PASSWORD: ${SALESFORCE_PLATFORM_PASSWORD}
       SALESFORCE_PLATFORM_SECURITY_TOKEN: ${SALESFORCE_PLATFORM_SECURITY_TOKEN}
+      AWS_PLATFORM_FILES_BUCKET: ${AWS_PLATFORM_FILES_BUCKET}
+      AWS_ACCESS_KEY_ID: ${AWS_ACCESS_KEY_ID}
+      AWS_SECRET_ACCESS_KEY: ${AWS_SECRET_ACCESS_KEY}
+       
+
       
     # Need privileged mode to change the sysctl flags in docker-compose/scripts/docker_compose_run.sh
     privileged: true
